+++ sass/custom/custom.scss
@@ -32,3 +32,8 @@
 @import 'components/user-card';
 @import 'components/utils';
 @import 'components/waveform';
+
+@import 'mobile/container';
+@import 'mobile/grid';
+@import 'mobile/header-search';
+@import 'mobile/padder';
