+++ sass/custom/mobile/container.scss
@@ -1,0 +1,5 @@
+@media only screen and (max-device-width: 480px) {
+    .container {
+        width: 100%;
+    }
+}
