+++ sass/custom/mobile/header-search.scss
@@ -1,0 +1,6 @@
+@media only screen and (max-device-width: 480px) {
+    .header-search {
+        display: none;
+    }
+
+}
