+++ sass/custom/mobile/padder.scss
@@ -1,0 +1,5 @@
+@media only screen and (max-device-width: 480px) {
+    .padder {
+        height: 0 !important;
+    }
+}
