+++ scripts/constants/ActionTypes.js
@@ -4,6 +4,7 @@ export const CHANGE_ACTIVE_SONG_ID = 'CHANGE_ACTIVE_SONG_ID';
 export const CHANGE_ACTIVE_USER_ID = 'CHANGE_ACTIVE_USER_ID';
 export const CHANGE_CURRENT_TIME = 'CHANGE_CURRENT_TIME';
 export const CHANGE_HEIGHT = 'CHANGE_HEIGHT';
+export const CHANGE_IS_MOBILE = 'CHANGE_IS_MOBILE';
 export const CHANGE_MODAL = 'CHANGE_MODAL';
 export const CHANGE_PATH = 'CHANGE_PATH';
 export const CHANGE_PLAYING_SONG = 'CHANGE_PLAYING_SONG';
