+++ scripts/containers/Mobile.js
@@ -1,0 +1,9 @@
+import React, {Component, PropTypes} from 'react';
+
+class Mobile extends Component {
+    render() {
+        return <div></div>;
+    }
+}
+
+export default Mobile;
