+++ scripts/reducers/index.js
@@ -1,6 +1,7 @@
 import {combineReducers} from 'redux';
 import authed from '../reducers/authed';
 import entities from '../reducers/entities';
+import environment from '../reducers/environment';
 import height from '../reducers/height';
 import modal from '../reducers/modal';
 import navigator from '../reducers/navigator';
@@ -10,6 +11,7 @@ import playlists from '../reducers/playlists';
 const rootReducer = combineReducers({
     authed,
     entities,
+    environment,
     height,
     modal,
     navigator,
